rag_tool_api/requirements.txt
fastapi
uvicorn
openai
pandas
google-cloud-storage
google-cloud-discoveryengine
protobuf
openpyxl 